backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Abi Raja <abimanyuraja@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.2"
uvicorn = "^0.27.0.post1"
websockets = "^12.0"
python-dotenv = "^1.0.1"
httpx = "^0.26.0"
openai = "^1.11.1"
sentry-sdk = {extras = ["fastapi"], version = "^1.40.2"}
asyncpg = "^0.29.0"
asyncpg-stubs = "^0.29.1"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
pre-commit = "^3.6.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pyright = "^1.1.350"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
